вынес игровые формы и обработчики к ним в отдельные модули.
